backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.93
botocore==1.34.93
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
cssselect==1.2.0
cssutils==2.10.2
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
emails==0.6
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi-filter==2.0.0
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.2.1
Mako==1.3.3
MarkupSafe==2.1.5
msgpack==1.0.8
passlib==1.7.4
premailer==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-filterset==2.0.0
starlette==0.37.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
